requirements.txt
Django==2.2.1
django-common-helpers==0.9.2
django-cron==0.5.1
django-google-maps==0.11.0
djangorestframework==3.9.3
flexdict==0.0.1a1
python-dateutil==2.8.0
pytz==2019.3
sentry-sdk==0.14.4
six==1.14.0
sqlparse==0.3.0
psycopg2==2.8.5